Add files via upload
For the mentor
Have you trained the trained the network with 20-30 number of epochs? I see that Siraj has trained the network for only one epoch which certainly is not enough. Can you put your codes on github or google drive and share the link?

As you can see, the train loss is always greater than the val_loss. Very confusing. Please help out. Thanks.